pyproject.toml
[tool.poetry]
name = "multicode-ai-bot"
version = "1.0.0"
description = "Multi-AI Telegram Bot - Access 8 AI coding assistants (Claude, Gemini, OpenAI, DeepSeek, Groq, Ollama, Blackbox, Windsurf) through Telegram"
authors = [
    "milhy545 <milhy777@gmail.com>",
    "Richard Atkinson <richardatk01@gmail.com>",
]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/milhy545/multicode-ai-bot"
repository = "https://github.com/milhy545/multicode-ai-bot"
documentation = "https://github.com/milhy545/multicode-ai-bot/blob/main/MULTI_AI_STATUS.md"
keywords = [
    "telegram", "bot", "ai", "claude", "gemini", "openai", "deepseek", "groq",
    "ollama", "multi-ai", "coding-assistant", "development", "remote", "llm"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Software Development :: Tools",
    "Topic :: System :: Shells",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^22.1"
structlog = "^25.4.0"
pydantic = "^2.11.5"
pydantic-settings = "^2.9.1"
python-dotenv = "^1.0.0"
aiofiles = "^24.1.0"
aiosqlite = "^0.21.0"
# AI Provider SDKs
anthropic = "^0.40.0"
claude-code-sdk = "^0.0.11"
google-generativeai = "^0.8.0"
aiohttp = "^3.9.0"

[tool.poetry.scripts]
multicode-bot = "src.main:run"
# Legacy alias for backwards compatibility
claude-telegram-bot = "src.main:run"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-asyncio = "^1.0.0"
pytest-cov = "^6.1.1"
pytest-mock = "^3.14.1"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"
mypy = "^1.16.0"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_errors = true

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503", "E501", "W291", "F401", "F541", "F841", "F811", "E722"]
